+++ src/factory.js
